+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Changed
 
+- Nuvlabox detail page - hide generate-new-api-key
 - Session - Error message in sign-in and sign-up pages not always
   visible [#778](https://github.com/nuvla/ui/issues/778)